:hammer: working on cum endpoint

+++ Server/U-Tracker API/src/main/kotlin/com/code_of_duty/utracker_api/data/models/Student.kt
@@ -2,6 +2,7 @@ package com.code_of_duty.utracker_api.data.models
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import jakarta.persistence.*
+import java.math.BigDecimal
 
 @Entity
 @Table(name="student")
@@ -16,7 +17,7 @@ data class Student(
     @Column(name = "image", nullable = true, columnDefinition = "varchar(255) default 'https://i.imgur.com/1qk9n3m.png'")
     var image: String? = null,
     @Column(name = "cum", nullable = false, columnDefinition = "float default 0.0")
-    var cum: Float = 0.0f,
+    var cum: BigDecimal = 0.0.toBigDecimal(),
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "degree_fk", referencedColumnName = "id", nullable = true)
     var degree: Degree? = null,
